Title Screen
Title screen with start message. Pressing start currently exits the
program gracefully. I think I learned why the only font available is
the 6X8 font. In the 8bkc-sdk project there is a config that seems
to only enable CONFIG_USE_FONT_6x8. The start message was too long
with spaces so I set the interfont spacing to zero before drawing
the string to the screen.

Todos on the title screen include showing a highscore and maybe the
git hash